+++ test/astar_square_euclidean_test.dart
@@ -11,32 +11,34 @@ void main() {
     astar.calculateGrid();
   });
   group('test AStarSquareGrid euclidean', () {
-    test('test last point', () {
-      final path = astar
-          .findPath(start: const Point(0, 0), end: const Point(3, 3))
+    test('test firs point is end point', () async {
+      final path = (await astar.findPath(
+              start: const Point(0, 0), end: const Point(3, 3)))
           .toPointList();
-      expect(path.last, const Point(3, 3)); // Check number of cols
+      expect(path.first, const Point(3, 3));
     });
 
-    test('test first point not start point', () {
-      final path = astar
-          .findPath(start: const Point(0, 0), end: const Point(3, 3))
+    test('test last point is start ', () async {
+      final path = (await astar.findPath(
+              start: const Point(0, 0), end: const Point(3, 3)))
           .toPointList();
-      expect(path.first != const Point(0, 0), true); // Check number of cols
+      expect(const [Point(0, 1), Point(1, 0), Point(1, 1)].contains(path.last),
+          true);
     });
 
-    test('test neighbors DiagonalMovement.manhattan return List.empty()', () {
-      final path = astar
-          .findPath(start: const Point(0, 0), end: const Point(0, 1))
+    test('test neighbors DiagonalMovement.manhattan return one item', () async {
+      final path = (await astar.findPath(
+              start: const Point(0, 0), end: const Point(0, 1)))
           .toPointList();
-      expect(path.isEmpty, true); // Check number of cols
+      expect(path.length == 1, true);
     });
 
-    test('test neighbors DiagonalMovement.manhattan diagonal return List.empty()', () {
-      final path = astar
-          .findPath(start: const Point(0, 0), end: const Point(1, 1))
+    test('test neighbors DiagonalMovement.manhattan diagonal return one item',
+        () async {
+      final path = (await astar.findPath(
+              start: const Point(0, 0), end: const Point(1, 1)))
           .toPointList();
-      expect(path.isEmpty, true); // Check number of cols
+      expect(path.length == 1, true);
     });
   });
 }