+++ test/blocked_path_test.dart
@@ -19,15 +19,15 @@ void main() {
     astar.calculateGrid();
   });
   group('test blocked path', () {
-    test('test if blocked ', () {
-      final path =
-          astar.findPath(start: const Point(0, 0), end: const Point(5, 5));
+    test('test if blocked ', () async {
+      final path = await astar.findPath(
+          start: const Point(0, 0), end: const Point(5, 5));
       expect(path.isEmpty, true); // Check number of cols
     });
 
-    test('test if not blocked ', () {
-      final path =
-          astar.findPath(start: const Point(0, 0), end: const Point(5, 2));
+    test('test if not blocked ', () async {
+      final path = await astar.findPath(
+          start: const Point(0, 0), end: const Point(5, 2));
       expect(path.isNotEmpty, true); // Check number of cols
     });
   });